refactor(agents): extract provider-config assembly from dispatch_plan god file
Second focused extraction from the agent_task_dispatch_plan god file
(after prompt_spec, #9357). Move the provider-config + secret-env cluster
into provider_config.rs: builds the executor provider_config object
(materialized refs + workspace metadata + request inputs), promotes
component contracts into it, and derives the secret-env allowlist.

Clean seam: the cluster is reached from the plan builder through 5
functions (dispatch_provider_config, dispatch_component_contracts,
dispatch_request_inputs, promote_component_contracts_to_provider_config,
dispatch_secret_env), re-exported pub(crate). Its only mod.rs dependency
is DispatchWorkspaceTarget (root + metadata fields), now pub(crate); it
also reuses prompt_spec::read_text_spec. External pub contract unchanged.

Behavior-preserving: mod.rs 1987 -> 1828 lines; provider_config.rs 181.
All 33 dispatch_plan tests pass.
